refactor: health indicator modules.
