+++ src/common/modules/health-indicators/elasticsearch.health/elasticsearch.health.indicator.ts
