+++ src/common/modules/health-indicators/elasticsearch.health/index.ts
